refactor: move generatePlotHTML from features/graph-enrichment.ts to presentation/plot.ts
Impact: 4 functions changed, 3 affected
